requirements.txt
altgraph==0.17.3
asttokens==2.2.1
async-generator==1.10
auto-py-to-exe==2.33.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
bottle==0.12.25
bottle-websocket==0.2.9
click==8.1.3
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dic==1.5.2b1
distro==1.8.0
dnspython==2.3.0
Eel==0.16.0
email-validator==2.0.0.post2
exceptiongroup==1.1.0
executing==1.2.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
imap-tools==1.0.0
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterthemes==0.20.0
keyboard==0.13.5
lesscpy==0.15.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
outcome==1.2.0
packaging==23.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
ply==3.11
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2023.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.0.2
selenium==4.8.0
six==1.16.0
SQLAlchemy==2.0.15
stack-data==0.6.2
tabula-py==2.6.0
tornado==6.3.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.4
typing_extensions==4.5.0
wcwidth==0.2.6
webdriver-manager==3.8.5
Werkzeug==2.3.4
whichcraft==0.6.1
wsproto==1.2.0
WTForms==3.0.1
xmltodict==0.13.0
zipp==3.15.0
zope.event==4.6
